Tournament module
The pokémon section its almost complete. You can add a complete team and
CRUD functions are working as intended. There is a small detail about
the edit (I can't get select2 to remember the old pokémon) and the
"add/remove from tournament" section of the module is missing.
Also there is a bug that allows you to add more than 6 pokémon for the
same tournament.