Huge database change
Now the users are stored in a different table public to all modules, not
just the tournament module. This is because we probably are going to
need to have users on more than one module (such as the players module)
so we can reuse the data.

I know its kind of ugly to change these things in these stages of the
development and I feel totally guilty =P.